brokk-acp-rust: extract permission gate into tool_loop/gate.rs (#3399)
Pure code-move: pulls PERMISSION_REQUEST_TIMEOUT, requires_per_call_prompt,
GateDecision, PureGateDecision, pure_gate_decision, SpawnedCx,
request_user_permission, the renamed consult function, and the gate
unit-test module out of the 800-line tool_loop.rs into its own submodule
alongside announce.rs.

The LLM turn loop in tool_loop::run is unchanged; the consult call site
becomes gate::consult(...). SpawnedCx is re-exported from tool_loop so
agent.rs's existing crate::tool_loop::SpawnedCx path keeps working.
Doc references in tools/sandbox.rs and tools/mod.rs updated to point at
tool_loop::gate::consult.

No behavior change. cargo fmt --check, cargo clippy --all-targets
-D warnings, and cargo test (111 passed) all clean.
